requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
inject==5.2.1
Mako==1.3.5
MarkupSafe==2.1.5
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
simple-settings==1.2.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.6
